requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.23
botocore==1.38.23
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.0.0
contourpy==1.3.2
cryptography==45.0.2
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
entrypoints==0.4
Faker==37.3.0
favicon==0.7.0
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
groq==0.25.0
h11==0.16.0
htbuilder==0.9.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-groq==0.3.2
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langchain_agents_pip==0.1.0
langchain_experimental>=0.0.30
langsmith==0.3.42
lxml==5.4.0
Markdown==3.8
markdownlit==0.0.7
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.7.1
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.40.0
numpy==1.24.3
openai==1.15.0
orjson==3.10.18
packaging==24.2
pandas==2.0.1
pdfminer.six==20250327
pdfplumber==0.10.0
pillow==10.0.0
platformdirs==4.3.8
plotly==5.14.1
prometheus_client==0.22.0
propcache==0.3.1
protobuf==5.29.4
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pymdown-extensions==10.15
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-py==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
s3transfer==0.13.0
setuptools==80.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.15.0
snowflake-snowpark-python==1.32.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.31.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.1
streamlit-faker==0.0.3
streamlit-image-coordinates==0.2.1
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
validators==0.35.0
wheel==0.45.1
yarl==1.20.0
zstandard==0.23.0
seaborn==0.12.2
wordcloud==1.9.2
PyMuPDF==1.23.19
tabula-py==2.9.0
opencv-python==4.8.0.76
Pillow==10.0.0
